Migrate from gist. Please read notes:
The only reason I used a gist for so long is because I'm a
commit perfectionist. With git I always want to `git add -p`
and craft single units of work, but a gist lets me allow
myself to say "fuck it, just paste the damned thing."

Since this project is optimized for my own joy I will
*endeavor* to allow myself to continue with fuck-it level
commits.